Fix #8 - This is a braking change for the calendar
Now querying /calendar returns a sorted list, to have the the timeline returned query with the new param timeline=true.

Calendar test were added, they sometime fail due to content availibility in the CMS, that will be change. The test suit will have to create content in the CMS (or Redis...)